agent/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
# OR for Claude:
# langchain-anthropic==0.1.0
requests==2.31.0
python-dotenv==1.0.0